### 2023-04-24 #5405 @editorialbot
[PRE REVIEW]: PxMCMC: A Python package for proximal Markov Chain Monte Carlo
**Submitting author:** @auggiemarignier (<a href="http://orcid.org/0000-0001-6778-1399">Augustin Marignier</a>)
**Repository:** https://github.com/auggiemarignier/pxmcmc
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v0.1.5
