### 2023-04-24 #5405 @auggiemarignier
Hi @arfon , no not a AAS paper but a [RASTI paper](https://academic.oup.com/rasti/article/2/1/20/6967146?login=false)
