### 2023-04-24 a936d7a
fix readthedocs config
