### 2023-06-14 #5405 @mstimberg
:wave: @lauralwatkins,  @tbrown122387, @arjunsavel, @farr would any of you be willing/available to review this submission for JOSS (The Journal of Open Source Software, a free, open-source, community driven and developer-friendly online journal, https://joss.theoj.org)?
