### 2023-06-20 #5405 @ejhigson
> 👋 @jeremiecoullon, @ejhigson, @SuperKam91 would any of you be willing/available to review this submission for JOSS (The Journal of Open Source Software, a free, open-source, community driven and developer-friendly online journal, https://joss.theoj.org/)?

I'm sorry I'm afraid I have left academia and do not have bandwidth to be a reviewer. Good luck with the submission though! :)
