### 2023-06-21 #5405 @mstimberg
:wave: @mattpitkin & @benjaminrose, would any of you be willing/available to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
