### 2023-06-22 #5582 @editorialbot
[REVIEW]: PxMCMC: A Python package for proximal Markov Chain Monte Carlo
**Submitting author:** @auggiemarignier (<a href="http://orcid.org/0000-0001-6778-1399">Augustin Marignier</a>)
**Repository:** https://github.com/auggiemarignier/pxmcmc
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v1.0.0
