### 2023-06-22 #5582 @mstimberg
👋🏼 @auggiemarignier, @jeremiecoullon, @mattpitkin, this is the review thread for the paper. All of our communications will happen here from now on.

at the top of a new comment in this thread.

There are additional guidelines in the [message at the start of this issue](https://github.com/openjournals/joss-reviews/issues/5582#issue-1769184675).

Please don't hesitate to ping me (@mstimberg) if you have any questions/concerns.
