### 2023-06-23 #5582 @mattpitkin
@auggiemarignier when looking at the examples scripts in https://github.com/auggiemarignier/pxmcmc/tree/main/experiments, there are no (or very few) comments in the codes. To make these far more useful as examples, could you please comment the major steps in each, in particular the set up of the sampler and the various inputs required.

Could you also please add come "Community guidelines" to the README/docs - see https://joss.readthedocs.io/en/latest/review_checklist.html#documentation.
