### 2023-06-28 1d9d6ba
add contributing statement to readme
