### 2023-06-28 4cfd76b
add comments to earthtopography example
