+++ experiments/earthtopography/plot.py
@@ -1,3 +1,7 @@
+"""
+Plots the summary maps from the MCMC in main.py
+"""
+
 import argparse
 import h5py
 import numpy as np
@@ -25,6 +29,7 @@ def filename(name):
     return f"{args.directory}/{name}{args.suffix}.png"
 
 
+# Load results file and extract global parameters
 file = h5py.File(args.datafile, "r")
 params = {attr: file.attrs[attr] for attr in file.attrs.keys()}
 L, B, J_min = params["L"], params["B"], params["J_min"]
@@ -33,24 +38,28 @@ try:
 except KeyError:
     setting = input("Specify setting:\t")
 nscales = pys2let.pys2let_j_max(B, L, J_min) - J_min + 1
-wvlttrans = SphericalWaveletTransform(
-    L,
-    B,
-    J_min,
-)
+
+# MCMC samples need to be transformed from wavelet to image space
+# so create helper transform instance
+wvlttrans = SphericalWaveletTransform(L, B, J_min)
 mw_shape = pyssht.sample_shape(L, Method="MW")
 
+# Plot the evolution of the posterior, likelihood and prior values
+# throughout the MCMC chain
 logpi = file["logposterior"][()]
 L2s = file["L2s"][()]
 L1s = file["priors"][()]
 evo = plotting.plot_evolution(logpi, L2s, L1s)
 evo.savefig(filename("evolution"))
 
+# Load original data, which in this case is also the ground truth
+# we are aiming to recover
 topo = hp.read_map("ETOPO1_Ice_hpx_256.fits", verbose=False, dtype=float,)
 truth = pyssht.inverse(pys2let.lm_hp2lm(map2alm(topo, L - 1), L), L, Reality=True) / 1000
 truthp = plotting.plot_map(truth, title="Truth")
 truthp.savefig(filename("truth"))
 
+# Get MAP estimate (i.e. highest posterior sample)
 MAP_idx = np.where(logpi == max(logpi))
 MAP_X = file["chain"][MAP_idx][0]
 if setting == "synthesis":
@@ -63,6 +72,7 @@ MAP = MAP.reshape(mw_shape).real
 maxapost = plotting.plot_map(MAP, title="Maximum a posetriori solution")
 maxapost.savefig(filename("MAP"))
 
+# Plot the difference with the truth
 diff = truth - MAP
 cbar_end = max([abs(np.min(diff)), np.max(diff)])
 diffp = plotting.plot_map(
@@ -74,23 +84,29 @@ diffp = plotting.plot_map(
 )
 diffp.savefig(filename("diff"))
 
+# Plot the values of the individual sampled parameters
 map_wvlt = plotting.plot_chain_sample(MAP_wvlt)
 map_wvlt.savefig(filename("MAP_wvlt"))
 
+# Make sure all the MCMC samples are in image space.
+# Drop the first args.burn samples.
 chain_pix = np.zeros(
     (file.attrs["nsamples"] - args.burn, pyssht.sample_length(L, Method="MW"))
 )
 for i, sample in enumerate(file["chain"][args.burn :]):
-    if setting == "synthesis":
+    if setting == "synthesis":  # sampled wavelet coeffecients so need to transform
         chain_pix[i] = wvlttrans.inverse(sample)
-    else:
+    else:  # sampled image space pixel values so no need to transform
         chain_pix[i] = np.copy(sample)
+
+# Calculate and plot the credible interval range
 ci_range = uncertainty.credible_interval_range(chain_pix).reshape(mw_shape)
 ci_map = plotting.plot_map(
     ci_range, title="95% credible interval range", cmap="viridis", vmin=0
 )
[…]
