+++ experiments/phasevel/plot.py
@@ -1,3 +1,7 @@
+"""
+Plot the results from main.py.
+"""
+
 import argparse
 import h5py
 import numpy as np
@@ -25,6 +29,7 @@ def filename(name, ext="png"):
     return f"{args.directory}/{name}{args.suffix}.{ext}"
 
 
+# Load results file, extract global parameters, setup wavelet transform
 file = h5py.File(args.datafile, "r")
 params = {attr: file.attrs[attr] for attr in file.attrs.keys()}
 L, B, J_min, setting = params["L"], params["B"], params["J_min"], params["setting"]
@@ -32,13 +37,14 @@ nscales = pys2let.pys2let_j_max(B, L, J_min) - J_min + 1
 wvlttrans = SphericalWaveletTransform(L, B, J_min,)
 mw_shape = pyssht.sample_shape(L, Method="MW")
 
+# Plot evolution of posterior, likelihood and prior
 logpi = file["logposterior"][()]
 L2s = file["L2s"][()]
 L1s = file["priors"][()]
 evo = plotting.plot_evolution(logpi, L2s, L1s)
 evo.savefig(filename("evolution"))
 
-
+# Plot MAP estimate
 MAP_idx = np.where(logpi == max(logpi))
 MAP_X = file["chain"][MAP_idx][0]
 if setting == "synthesis":
@@ -53,6 +59,7 @@ maxapost = plotting.plot_map(
 )
 maxapost.savefig(filename("MAP"))
 
+# Load the ground truth image and plot the difference
 truth = np.load("ekstrom/ekstrom28.npy")
 diff = truth - MAP
 diff_perc = 100 * diff / np.max(abs(truth))
@@ -65,7 +72,7 @@ diffp.savefig(filename("diff"))
 map_wvlt = plotting.plot_chain_sample(MAP_wvlt)
 map_wvlt.savefig(filename("MAP_wvlt"))
 
-
+# Get all samples into image space
 chain_pix = np.zeros(
     (file.attrs["nsamples"] - args.burn, pyssht.sample_length(L, Method="MW"))
 )
@@ -74,6 +81,8 @@ for i, sample in enumerate(file["chain"][args.burn :]):
         chain_pix[i] = wvlttrans.inverse(sample)
     else:
         chain_pix[i] = np.copy(sample)
+
+# Calculate and plot the credible interval range
 alpha = 0.01
 quantiles = np.quantile(chain_pix, (alpha / 2, 1 - alpha / 2), axis=0)
 in_ci = (truth.flatten() >= quantiles[0]) & (truth.flatten() <= quantiles[1])
@@ -83,6 +92,7 @@ ci_map = plotting.plot_map(
 )
 ci_map.savefig(filename("ci_map"))
 
+# Calculate and plot the credible interval range for each wavelet scale
 wav_ci_ranges = uncertainty.wavelet_credible_interval_range(
     file["chain"][args.burn :], L, B, J_min
 )
@@ -100,12 +110,14 @@ for i, wav_ci_range in enumerate(wav_ci_ranges):
     )
     wav_ci_map.savefig(filename(f"ci_map_scale{i}"))
 
+# Calculate and plot the mean
 mean = np.mean(chain_pix, axis=0).reshape(mw_shape)
 mean_map = plotting.plot_map(
     np.ascontiguousarray(mean), title="Mean solution", cmap="seismic_r", centre0=True
 )
 mean_map.savefig(filename("mean"))
 
+# Get the mean at at each wavelet scale
 figs = plotting.plot_wavelet_maps(
     mean, L, B, J_min, title="Mean solution", cmap="seismic_r", centre0=True
 )
[…]
