+++ experiments/weaklensing/plot.py
@@ -1,3 +1,7 @@
+"""
+Plot the results from main.py.
+"""
+
 import argparse
 import h5py
 import numpy as np
@@ -27,14 +31,16 @@ def filename(name, ext="png"):
     return f"{args.directory}/{name}{args.suffix}.{ext}"
 
 
+# Load results file and extract global parameters
 file = h5py.File(args.datafile, "r")
 params = {attr: file.attrs[attr] for attr in file.attrs.keys()}
 L, B, J_min, setting = params["L"], params["B"], params["J_min"], params["setting"]
 nscales = pys2let.pys2let_j_max(B, L, J_min) - J_min + 1
 wvlttrans = SphericalWaveletTransform(L, B, J_min,)
 mw_shape = pyssht.sample_shape(L, Method="MW")
-oversample = L < 256
+oversample = L < 256  # get smoother images by padding spectrum with zeros
 
+# Plot evolution of posterior, likelihood and prior probabilities
 logpi = file["logposterior"][()]
 L2s = file["L2s"][()]
 L1s = file["priors"][()]
@@ -43,6 +49,7 @@ evo.savefig(filename("evolution"))
 
 lmax = L - 1
 nside = 3 * lmax - 1
+# Get the ground truth for comparison
 with warnings.catch_warnings():
     warnings.simplefilter("ignore")
     kappa = hp.read_map("takahasi_4096_000_zs16_kappa.fits")
@@ -51,6 +58,9 @@ with warnings.catch_warnings():
     kappa_s = hp.smoothing(kappa_bl, sigma=sigma)
 truth = pyssht.inverse(pys2let.lm_hp2lm(hp.map2alm(kappa_s, lmax), L), L)
 
+
+# Need 1 mask with the same bandlimit as the results
+# and one at the higher bandlimit that is used for plotting
 if args.no_mask:
     mask = np.ones(pyssht.sample_shape(L)).astype(bool)
     highL_mask = (
@@ -60,8 +70,7 @@ else:
     mask = build_mask(L, size=20).astype(bool)
     highL_mask = build_mask(256, size=20).astype(bool) if oversample else np.copy(mask)
 
-wl = WeakLensing(L, mask)
-
+# Get and plot the MAP estimate
 MAP_idx = np.where(logpi == max(logpi))
 MAP_X = file["chain"][MAP_idx][0]
 if setting == "synthesis":
@@ -81,6 +90,7 @@ maxapost = plotting.plot_map(
 )
 maxapost.savefig(filename("MAP"))
 
+# Plot the difference with the ground truth
 diff = np.ascontiguousarray(truth - MAP).real
 diff_perc = 100 * diff / np.max(abs(truth))
 cbar_end = min(max([abs(np.min(diff * mask)), np.max(diff * mask)]), 100)
@@ -98,7 +108,7 @@ diffp.savefig(filename("diff"))
 map_wvlt = plotting.plot_chain_sample(MAP_wvlt)
 map_wvlt.savefig(filename("MAP_wvlt"))
 
-
+# Get everythin in image space
 chain_pix = np.zeros(
     (file.attrs["nsamples"] - args.burn, pyssht.sample_length(L, Method="MW")),
     dtype=complex,
@@ -108,6 +118,8 @@ for i, sample in enumerate(file["chain"][args.burn :]):
         chain_pix[i] = wvlttrans.inverse(sample)
     else:
         chain_pix[i] = np.copy(sample)
+
+# Get and plot the uncertainty
 ci_range = uncertainty.credible_interval_range(chain_pix).reshape(mw_shape)
 ci_map = plotting.plot_map(
[…]
