### 2023-06-29 #5582 @auggiemarignier
@mattpitkin thanks for your comments.  I've opened a [PR](https://github.com/auggiemarignier/pxmcmc/pull/15) in the main repo where I'll implement the various changes suggested during the review.

I've added comments throughout the example scripts (data still needs to be added so that they run), and added the community guidelines to README and docs CONTRIBUTING page.
