### 2023-07-05 07e3aed
add synthetic data and ground truth for phasevel example
