### 2023-07-05 2bc81bd
add help to command line args
