+++ experiments/weaklensing/plot.py
@@ -18,12 +18,27 @@ from pxmcmc.utils import snr, norm, build_mask
 
 
 parser = argparse.ArgumentParser()
-parser.add_argument("datafile", type=str)
-parser.add_argument("directory", type=str)
-parser.add_argument("--suffix", type=str, default="")
-parser.add_argument("--burn", type=int, default=1000)
-parser.add_argument("--save_npy", action="store_true")
-parser.add_argument("--no-mask", action="store_true")
+parser.add_argument(
+    "datafile", type=str, help="Path to .hdf5 file with pxmcmc results."
+)
+parser.add_argument("directory", type=str, help="Directory in which to save plots.")
+parser.add_argument(
+    "--suffix", type=str, default="", help="Optional suffix to output filenames."
+)
+parser.add_argument(
+    "--burn",
+    type=int,
+    default=1000,
+    help="Ignore the first <burn> MCMC samples.  Default 100.",
+)
+parser.add_argument(
+    "--save_npy",
+    action="store_true",
+    help="Also save the output summary maps as .npy files",
+)
+parser.add_argument(
+    "--no-mask", action="store_true", help="Reveal what is behind the Euclid mask"
+)
 args = parser.parse_args()
 
 
