### 2023-07-05 cd978a5
add simple readmes for the examples
