### 2023-07-06 #5582 @jeremiecoullon
@auggiemarignier: I'm having trouble instaling the package from source. I opened an issue on github describing my setup and the error I get.
I wonder if you have thoughts on this?

Thanks!
