### 2023-07-07 auggiemarignier/pxmcmc#17 @mattpitkin
Internal version string does not match the package version
The `__version__` string in `__init__.py` (0.1.1) does not match the current package version (0.1.5). You could update this manually and create a new release with the correct version in it, or you might want to switch to an automated version management system such as versioneer or setuptools_scm.
