### 2023-07-07 #5582 @mattpitkin
I've opened an issue about the internal package version string not matching the current package version https://github.com/auggiemarignier/pxmcmc/issues/17.
