### 2023-07-08 f3cdba2
add versioneer config
