+++ pyproject.toml
@@ -32,9 +32,17 @@ flake8 = "^3.7.9"
 pytest-cases = "^3.6.13"
 
 [build-system]
-requires = ["poetry>=0.12"]
+requires = ["poetry>=0.12", "versioneer[toml]==0.29"]
 build-backend = "poetry.masonry.api"
 
 [tool.poetry.extras]
 docs = ["sphinx", "sphinx-rtd-theme"]
 cartopy = ["Cartopy"]
+
+[tool.versioneer]
+VCS = "git"
+style = "pep440"
+versionfile_source = "pxmcmc/_version.py"
+versionfile_build = ""
+tag_prefix = "v"
+parentdir_prefix = ""
