+++ pxmcmc/__init__.py
@@ -1,1 +1,4 @@
 __version__ = '0.1.1'
+
+from . import _version
+__version__ = _version.get_versions()['version']
