### 2023-07-08 auggiemarignier/pxmcmc#17 @auggiemarignier
@mattpitkin using `importlib.metadata.version()` in `__init__.py` and version bumping with `poetry version`

https://github.com/auggiemarignier/pxmcmc/pull/15/commits/2acd60b0d339bb755cf8b3fe3f167c1b526ad8ce
