+++ paper/paper.bib
@@ -29,7 +29,8 @@
   year    = {2006},
   volume  = {52},
   number  = {4},
-  pages   = {1289-1306}
+  pages   = {1289-1306},
+  doi     = {10.1109/TIT.2006.871582},
 }
 
 @article{Foreman-Mackey2013,
