### 2023-07-13 #5582 @mattpitkin
In terms of "State of the field", you might want to mention this package https://github.com/astro-informatics/proxnest, which uses the proximal method for nested sampling.
