+++ paper/paper.md
@@ -42,8 +42,8 @@ The use of proximal operators in MCMC was first proposed by @Pereyra2016, modify
 
 MCMC methods already have popular implementations.
 For example, gradient-based Hamiltonian Monte Carlo is implemented in `STAN` [@Stan], and emcee [@Foreman-Mackey2013] is a Python implementation of the affine-invariant ensemble sampler MCMC [@Goodman2010] popular in the astrophysics community.
-To the author's knowledge, however, there exists no Python implementation of proximal MCMC readily available.
-PxMCMC is a Python package implementing proximal algorithms from @Pereyra2016 and @Pereyra2020.
+To the author's knowledge, however, there exists no Python implementation of proximal MCMC readily available other than a recent implementation of proximal nested sampling [@Cai2021] who's primary aim is calculating the marginal likelihood.
+PxMCMC is a Python package implementing proximal algorithms from @Pereyra2016 and @Pereyra2020 for efficiently obtaining posterior samples form high-dimensional parameter spaces.
 The class-based API abstracts out the main components of MCMC into interoperable classes, thereby allowing users to implement their own forward models (physical model) and priors, and even their own MCMC sampler if desired.
 Originally developed to solve inverse imaging problems defined on spherical domains [@Marignier2023; @Marignier2023thesis], the package provides priors to promote sparsity in a spherical wavelet domain using transforms from the `S2LET` package [@Leistedt2013].
 Examples provided in the package include a common problem in global seismic tomography and a full-sky cosmological mass-mapping problem, the details of which can be found in @Marignier2023 and @Marignier2023thesis.
