+++ paper/paper.bib
@@ -11,6 +11,14 @@
   year          = {2018}
 }
 
+@article{Cai2021,
+  author  = {Cai, Xiaohao and McEwen, Jason~D. and Pereyra, Marcelo},
+  title   = {"High-dimensional Bayesian model selection by proximal nested sampling"},
+  journal = {ArXiv},
+  eprint  = {arXiv:2106.03646},
+  year    = {2021}
+}
+
 @article{Candes2011,
   title   = {Compressed sensing with coherent and redundant dictionaries},
   volume  = {31},
@@ -30,36 +38,36 @@
   volume  = {52},
   number  = {4},
   pages   = {1289-1306},
-  doi     = {10.1109/TIT.2006.871582},
+  doi     = {10.1109/TIT.2006.871582}
 }
 
 @article{Foreman-Mackey2013,
-	doi = {10.1086/670067},
-	url = {https://doi.org/10.1086%2F670067},
-	year = 2013,
-	month = {mar},
-	publisher = {{IOP} Publishing},
-	volume = {125},
-	number = {925},
-	pages = {306--312},
-	author = {Daniel Foreman-Mackey and David W. Hogg and Dustin Lang and Jonathan Goodman},
-	title = {$\less$tt$\greater$emcee$\less$/tt$\greater$: The {MCMC} Hammer},
-	journal = {Publications of the Astronomical Society of the Pacific}
+  doi       = {10.1086/670067},
+  url       = {https://doi.org/10.1086%2F670067},
+  year      = 2013,
+  month     = {mar},
+  publisher = {{IOP} Publishing},
+  volume    = {125},
+  number    = {925},
+  pages     = {306--312},
+  author    = {Daniel Foreman-Mackey and David W. Hogg and Dustin Lang and Jonathan Goodman},
+  title     = {$\less$tt$\greater$emcee$\less$/tt$\greater$: The {MCMC} Hammer},
+  journal   = {Publications of the Astronomical Society of the Pacific}
 }
 
-@ARTICLE{Goodman2010,
-       author = {{Goodman}, Jonathan and {Weare}, Jonathan},
-        title = "{Ensemble samplers with affine invariance}",
-      journal = {Communications in Applied Mathematics and Computational Science},
-     keywords = {Markov chain Monte Carlo, affine invariance, ensemble samplers},
-         year = 2010,
-        month = jan,
-       volume = {5},
-       number = {1},
-        pages = {65-80},
-          doi = {10.2140/camcos.2010.5.65},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2010CAMCS...5...65G},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+@article{Goodman2010,
+  author   = {{Goodman}, Jonathan and {Weare}, Jonathan},
+  title    = {{Ensemble samplers with affine invariance}},
+  journal  = {Communications in Applied Mathematics and Computational Science},
+  keywords = {Markov chain Monte Carlo, affine invariance, ensemble samplers},
+  year     = 2010,
+  month    = jan,
+  volume   = {5},
+  number   = {1},
+  pages    = {65-80},
+  doi      = {10.2140/camcos.2010.5.65},
+  adsurl   = {https://ui.adsabs.harvard.edu/abs/2010CAMCS...5...65G},
+  adsnote  = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
 @article{Leistedt2013,
@@ -75,25 +83,25 @@
 }
 
 @article{Marignier2023,
-    author = {Marignier, Augustin and McEwen, Jason D and Ferreira, Ana M G and Kitching, Thomas D},
[…]
