+++ paper/paper.bib
@@ -16,7 +16,8 @@
   title   = {"High-dimensional Bayesian model selection by proximal nested sampling"},
   journal = {ArXiv},
   eprint  = {arXiv:2106.03646},
-  year    = {2021}
+  year    = {2021},
+  doi     = {arXiv:2106.03646}
 }
 
 @article{Candes2011,
