### 2023-07-16 #5582 @auggiemarignier
@mattpitkin just chasing up on the last few things on your checklist
 - *Functionality - Functionality* : is there anything I need to change for this?  I recommend simply running the Earth topography example.
 - *Documentation - Statement of need* : see the [docs build for the PR](https://pxmcmc--15.org.readthedocs.build/en/15/)
 - *Software Paper - State of the field* : I've added the proxnest package as you suggested (see latest proof)
 - *Software Paper - References* : I believe everything is there and the syntax is correct.

Let me know if there's anything else!
