### 2023-07-24 #5582 @jeremiecoullon
I finished my review: I think it's very good!

The examples are nice, the code is well structured, and the UX (ie: running the quickstart, interacting with the repo & documentation) is good.

By the way, I suggested some [small edits](https://github.com/auggiemarignier/pxmcmc/pull/15#issuecomment-1647461373) which I think would further reduce friction in getting a first example up and running.

But this is all great! Congrats on this great package! 🎉 ✅
