### 2023-07-25 88761ad
add examples and citation to main README.md
