### 2023-07-25 0316790
remove pyssht, pys2let and astropy from deps
