+++ pyproject.toml
@@ -13,13 +13,9 @@ readme = "README.md"
 python = ">=3.8,<4.0"
 numpy = "^1.21.1"
 matplotlib = "^3.4.2"
-pyssht = "^1.4.0"
-pys2let = "^2.2.3"
-greatcirclepaths = "^1.1.0"
 h5py = "^3.3.0"
 Cartopy = {version = "^0.19.0", optional = true}
 sphinx = {version = "^4", optional = true}
-astropy = "^5.0.4"
 sphinx-rtd-theme = {version = "^1.0.0", optional = true}
 scipy = "^1.9.2"
 
