### 2023-07-25 a2db1dd
replace some utility functions from ssht and s2let
