+++ tests/test_utils.py
@@ -3,6 +3,7 @@ import numpy as np
 import pytest
 import pys2let
 import pyssht
+from s2fft import sampling
 
 
 def test_flattenmlm():
@@ -87,14 +88,14 @@ def test_s2_integrate(L):
     flm = np.zeros((L * L), dtype=complex)
     for el in range(L):
         m = 0
-        ind = pyssht.elm2ind(el, m)
+        ind = sampling.elm2ind(el, m)
         flm[ind] = np.random.randn()
         for m in range(1, el + 1):
-            ind_pm = pyssht.elm2ind(el, m)
-            ind_nm = pyssht.elm2ind(el, -m)
+            ind_pm = sampling.elm2ind(el, m)
+            ind_nm = sampling.elm2ind(el, -m)
             flm[ind_pm] = np.random.randn() + 1j * np.random.randn()
             flm[ind_nm] = (-1) ** m * np.conj(flm[ind_pm])
     I0 = flm[0] * np.sqrt(4 * np.pi)
-    f = pyssht.inverse(flm, L, Method="MW", Reality=True).flatten()
+    f = pyssht.inverse(flm, L, Reality=True).flatten()
 
     assert np.isclose(I0, utils.s2_integrate(f, L))
