+++ pxmcmc/prior.py
@@ -1,6 +1,6 @@
 import numpy as np
-import pys2let
-import pyssht
+import s2wav
+import s2fft
 from s2fft import sampling
 from s2wav.utils.shapes import j_max
 
