### 2023-07-25 919b0b2
remove healpy imports for now (not in current env)
