+++ pxmcmc/plotting.py
@@ -1,5 +1,5 @@
 import numpy as np
-import healpy as hp
+
 import matplotlib.pyplot as plt
 from matplotlib import cm
 import copy
