+++ pyproject.toml
@@ -19,6 +19,7 @@ sphinx = {version = "^4", optional = true}
 sphinx-rtd-theme = {version = "^1.0.0", optional = true}
 scipy = "^1.9.2"
 s2wav = {git = "https://github.com/astro-informatics/s2wav.git"}
+jaxlib = "^0.4.13"
 
 [tool.poetry.dev-dependencies]
 pytest = "^7.0"
