### 2023-07-25 c9915ed
remove astropy for now (not in current env)
