### 2023-07-25 #5582 @mstimberg
Here are a few suggestions for minor edits to the paper:
- l.33–34: I think this should use `[e.g. @Cai2018; @Price2019; @Marignier2023]` to avoid having the brackets-within-brackets citations.
- l.41: "who's" → "whose"
- l.43: "form" → "for" (or "from"?)

References:
- Cai et al. (2021): The DOI should be `10.48550/arXiv.2106.03646`
- Foreman-Mackey (2013): In the bib file `$\less$tt$\greater$emcee$\less$/tt$\greater$` → `emcee`
- Marignier, A. (2023): maybe add a URL?
- Please remove `type= {Journal Article}` from Pereyra2020 and Price2019
- In Pereyra2020: please use `title   = {Accelerating Proximal {Markov Chain Monte Carlo} by Using an Explicit Stabilized Method}` (or fix the capitalization manually to use sentence case).
