+++ paper/paper.bib
@@ -15,9 +15,9 @@
   author  = {Cai, Xiaohao and McEwen, Jason~D. and Pereyra, Marcelo},
   title   = {"High-dimensional Bayesian model selection by proximal nested sampling"},
   journal = {ArXiv},
-  eprint  = {arXiv:2106.03646},
+  eprint  = {10.48550/arXiv.2106.03646},
   year    = {2021},
-  doi     = {arXiv:2106.03646}
+  doi     = {10.48550/arXiv.2106.03646}
 }
 
 @article{Candes2011,
@@ -52,7 +52,7 @@
   number    = {925},
   pages     = {306--312},
   author    = {Daniel Foreman-Mackey and David W. Hogg and Dustin Lang and Jonathan Goodman},
-  title     = {$\less$tt$\greater$emcee$\less$/tt$\greater$: The {MCMC} Hammer},
+  title     = {emcee: The {MCMC} Hammer},
   journal   = {Publications of the Astronomical Society of the Pacific}
 }
 
@@ -102,7 +102,8 @@
   title  = {From Dark Matter to the Earth's Deep Interior: There and Back Again},
   author = {Marignier, Augustin},
   year   = {2023},
-  school = {UCL (University College London)}
+  school = {UCL (University College London)},
+  url    = {https://discovery.ucl.ac.uk/id/eprint/10162902/}
 }
 
 @article{Moreau1962,
@@ -145,14 +146,13 @@
 
 @article{Pereyra2020,
   author  = {Pereyra, Marcelo and Mieles, Luis Vargas and Zygalakis, Konstantinos C.},
-  title   = {{Accelerating Proximal Markov Chain Monte Carlo by Using an Explicit Stabilized Method}},
+  title   = {{Accelerating Proximal Markov Chain Monte Carlo by using an explicit stabilized method}},
   journal = {{SIAM} Journal on Imaging Sciences},
   volume  = {13},
   number  = {2},
   pages   = {905-935},
   doi     = {10.1137/19m1283719},
-  year    = {2020},
-  type    = {Journal Article}
+  year    = {2020}
 }
 
 @article{Price2019,
@@ -163,8 +163,7 @@
   number  = {1},
   pages   = {394-404},
   doi     = {10.1093/mnras/stz3453},
-  year    = {2020},
-  type    = {Journal Article}
+  year    = {2020}
 }
 
 @misc{Stan,
