### 2023-07-26 #5582 @auggiemarignier
@mstimberg I've addressed your final comments.  I've released `v1.0.0` and the zenodo DOI is `10.5281/zenodo.8185139`
